added gun Pick Up Script, added a spider and skeleton enemies, created a sword, doors and enemies prefabs

+++ Assets/Scripts/gunStats.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu]
 public class gunStats : ScriptableObject
 {
-    public int shootRate;
+    public float shootRate;
     public int shootDistance;
     public int shootDamage;
     public int ammoCount;
